point normalization

+++ drawing_commons/src/models.rs
@@ -108,11 +108,11 @@ pub fn get_feature_names() -> [String; 2] {
 #[derive(Clone, Deserialize, Serialize)]
 pub struct SampleWithFeatures {
     pub label: String,
-    pub point: Vec<usize>,
+    pub point: Vec<f64>,
 }
 
 impl SampleWithFeatures {
-    pub fn create(label: String, point: Vec<usize>) -> Self {
+    pub fn create(label: String, point: Vec<f64>) -> Self {
         Self { label, point }
     }
 }
